feat: Clean map when destroy socket
When a socket has been closed in kernel, the corresponding connection
info in user space should be cleaned in order to avoid memory leaking.

Then, in user space, use another map to keep pid and fd for the socket
with socket pointer in kernel space as key. It is because the tuple info
in the socket is incorrect when destroy it.

e.g.

```
2024-12-13T13:23:10Z DBG AddConn success fd=5 pid=134386 tuple=192.168.241.133:35348-172.217.194.113:443
2024-12-13T13:23:10Z DBG DestroyConn success fd=5 pid=134386 tuple=192.168.241.133:35348-172.217.194.113:443
2024-12-13T13:23:16Z DBG AddConn success fd=4 pid=134404 tuple=192.168.241.1:65063-192.168.241.133:8080
2024-12-13T13:23:16Z DBG DestroyConn success fd=4 pid=134404 tuple=192.168.241.1:65063-192.168.241.133:8080
```

Signed-off-by: Leon Hwang <hffilwlqm@gmail.com>